feat: introduce services and controller for data migration and TMDB import functionality.
